Skip bookmarks for missing items instead of failing
Currently if a bookmark exists for an item that does not exist in the
db, either because it has been removed or it never did, the user is
prevented from retrieving any bookmarks. We can be in this state if
media has moved or been deleted, but also because the createBookmark
endpoint does not verify the item id provided.

This patch makes the getBookmarks endpoint ignore bookmarks for missing
items.

Fixes: https://github.com/sentriz/gonic/issues/578

Signed-off-by: Eric B Munson <eric@munsonfam.org>